pyproject.toml
[tool.poetry]
name = "msg-broadcast"
version = "0.1.0"
description = ""
authors = ["hjmr <2620990+hjmr@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
line-bot-sdk = "^3.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
